Fix image of a tile still got displayed even after the piece has been moved.
The previous implementation didn't check the existence of a piece when resetting the image of the initial tile it's start dragged from. When a player dragged a piece outside of the board and then release a left click, the `Tile_MouseLeftButtonUp` event was not fired, so when the cursor moved across the board the piece was still in a dragging state. Normally this would not cause any issue unless a player click a tile which was related to a certain tile action (tile that was displaying a hint). When that happen, the `ChessGame` will invoke a `PieceMoved` event, telling the UI that a piece has moved from one location to another, at the same time, the UI will reset the starting tile to it's initial image, this will lead to a visual bug as if there was a duplicate of two pieces. To address this issue, the `StopDraggingTile()` method now check whether the tile is currently occupied before resetting back a tile to it's initial state.